example/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
backcall==0.1.0           # via ipython
click==6.7                # via pip-tools
decorator==4.3.0          # via ipython, traitlets
django-autofixture==0.12.1
django==1.11.13
first==2.0.1              # via pip-tools
flake8==3.5.0
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==6.4.0
jedi==0.12.0              # via ipython
mccabe==0.6.1             # via flake8
parso==0.2.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==2.0.2
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.5.2         # via pexpect
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0           # via flake8
pygments==2.2.0           # via ipython
pytz==2018.4              # via django
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via pip-tools, prompt-toolkit, traitlets
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
